Renaming in Metadata Service Tests - dependencies updated
