+++ tests/metadata-service-tests/Gopkg.toml
@@ -31,8 +31,7 @@
 
 [[constraint]]
   name = "github.com/kyma-project/kyma"
-  source = "github.com/Szymongib/kyma"
-  revision = "4883555e8bc822460b6e8bd43aee20e9bcb37884"
+  revision = "d6ba814ef8b42b4c7c65f478e51f46f01387e1df"
 
 [prune]
   go-tests = true
